+++ Landscape/game_mod/adv_mt.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <nh3api/core/map.hpp>
+#include "types.hpp"
+
+
+bool __fastcall IsTerrainType(const NewfullMap& map, const CObject& obj, int width, TerrainType8 terType);
+
+TerrainType8 __fastcall GetTerrainType(const NewfullMap& map, const CObject& obj, int width);
+
+void __fastcall MarkMagicTerrain(NewfullMap& map);